requirements.txt
PyQt5==5.15.10
loguru
python-dotenv
pytest
black
ruff
